@@ -3,9 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
     "@githubprimer/octicons-react": "^8.5.0",
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
+    "@mui/icons-material": "^5.1.0",
+    "@mui/material": "^5.1.0",
     "@types/classnames": "^2.3.1",
     "@types/react": "^17.0.8",
     "@types/react-dom": "^17.0.5",
@@ -33,23 +35,24 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.26.0",
-    "@typescript-eslint/parser": "^4.26.0",
-    "eslint": "^7.27.0",
+    "@emotion/eslint-plugin": "^11.5.0",
+    "@typescript-eslint/eslint-plugin": "^5.3.1",
+    "@typescript-eslint/parser": "^5.3.1",
+    "eslint": "^8.2.0",
     "eslint-plugin-react": "^7.24.0",
-    "gatsby": "^3.14.1",
+    "gatsby": "^4.1.4",
     "gatsby-image": "^3.11.0",
-    "gatsby-plugin-google-analytics": "^3.14.0",
-    "gatsby-plugin-manifest": "^3.14.0",
-    "gatsby-plugin-material-ui": "^3.0.1",
-    "gatsby-plugin-no-sourcemaps": "^3.14.0",
-    "gatsby-plugin-offline": "^4.14.0",
-    "gatsby-plugin-react-helmet": "^4.14.0",
-    "gatsby-plugin-sharp": "^3.14.0",
-    "gatsby-plugin-typescript": "^3.14.0",
-    "gatsby-source-filesystem": "^3.14.0",
-    "gatsby-transformer-remark": "^4.11.0",
-    "gatsby-transformer-sharp": "^3.14.0",
+    "gatsby-plugin-emotion": "^7.1.0",
+    "gatsby-plugin-google-analytics": "^4.1.0",
+    "gatsby-plugin-manifest": "^4.1.3",
+    "gatsby-plugin-no-sourcemaps": "^4.1.0",
+    "gatsby-plugin-offline": "^5.1.3",
+    "gatsby-plugin-react-helmet": "^5.1.0",
+    "gatsby-plugin-sharp": "^4.1.3",
+    "gatsby-plugin-typescript": "^4.1.2",
+    "gatsby-source-filesystem": "^4.1.2",
+    "gatsby-transformer-remark": "^5.1.3",
+    "gatsby-transformer-sharp": "^4.1.0",
     "prettier": "^2.3.0",
     "typescript": "^4.4.3"
   },
