@@ -1,11 +1,11 @@
-import React, { HTMLProps } from "react";
+import { HTMLProps } from "react";
 import { Link } from "gatsby";
-import CardMedia from "@material-ui/core/CardMedia";
-import CardContent from "@material-ui/core/CardContent";
-import CardActions from "@material-ui/core/CardActions";
-import Typography from "@material-ui/core/Typography";
-import Button from "@material-ui/core/Button";
-import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
+import { css } from "@emotion/react";
+import CardMedia from "@mui/material/CardMedia";
+import CardContent from "@mui/material/CardContent";
+import CardActions from "@mui/material/CardActions";
+import Typography from "@mui/material/Typography";
+import Button from "@mui/material/Button";
 
 export interface ProjectCardProps extends HTMLProps<HTMLDivElement> {
   // project: Project;
@@ -17,55 +17,55 @@ export interface ProjectCardProps extends HTMLProps<HTMLDivElement> {
   repo?: string;
 }
 
-const styles = (theme: Theme) =>
-  createStyles({
-    card: () => ({
-      // Height and width
-      height: "100%",
-      // Flex
-      display: "flex",
-      flexDirection: "column",
-      // Color
-      backgroundColor: theme.palette.background.paper,
-      // border
-      border: `1px solid ${theme.palette.grey[300]}`,
-      borderRadius: theme.shape.borderRadius,
-    }),
-    cardTitle: {
-      paddingBottom: theme.spacing(),
-    },
-    cardDescription: {
-      flexGrow: 1,
-      paddingTop: 0,
-      height: "auto",
-      overflowY: "auto",
-    },
-    cardActions: {
-      flexShrink: 0,
-    },
-  });
-
-const useStyles = makeStyles(styles);
-
 /**
  * Display a project using a card
  */
 function ProjectCard(props: ProjectCardProps) {
-  const classes = useStyles(props);
   const { title, role, image, learnMore, repo, overview, ...others } = props;
 
   return (
-    <div className={classes.card} {...others}>
+    <div
+      css={(theme) => css`
+        // Height and width
+        height: 100%;
+        // Flex
+        display: flex;
+        flex-direction: column;
+        // Color
+        background-color: ${theme.palette.background.paper};
+        // border
+        border: 1px solid ${theme.palette.grey[300]};
+        border-radius: ${theme.shape.borderRadius}px;
+      `}
+      {...others}
+    >
       {!image ? null : <CardMedia image={image}></CardMedia>}
       {/* project titles, fixed size */}
-      <CardContent className={classes.cardTitle}>
+      <CardContent
+        css={(theme) => css`
+          padding-bottom: ${theme.spacing()}px;
+        `}
+      >
         <Typography variant="subtitle1">{title}</Typography>
         <Typography variant="subtitle2">{role}</Typography>
       </CardContent>
       {/* project descriptions, automatic take up all reamining sizes */}
-      <CardContent className={classes.cardDescription}>{overview}</CardContent>
+      <CardContent
+        css={css`
+          flex-grow: 1;
+          padding-top: 0;
+          height: auto;
+          overflowy: auto;
+        `}
+      >
+        {overview}
+      </CardContent>
       {/* card actions, fixed size */}
-      <CardActions className={classes.cardActions}>
+      <CardActions
+        css={css`
+          flex-shrink: 0;
+        `}
+      >
         {!learnMore ? null : (
           <Button size="small" color="primary" component={Link} to={learnMore}>
             Learn More
