@@ -1,8 +1,7 @@
-import React from "react";
 import { graphql, StaticQuery } from "gatsby";
 import Helmet from "react-helmet";
-import Container from "@material-ui/core/Container";
-import Grid from "@material-ui/core/Grid";
+import Container from "@mui/material/Container";
+import Grid from "@mui/material/Grid";
 import Layout from "../../components/Layout";
 import TopBar from "../../components/TopSpace";
 import ProjectCard from "../../components/ProjectCard";
