@@ -1,18 +1,13 @@
-import React from "react";
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
-
-const styles = (theme: Theme) =>
-  createStyles({
-    container: {
-      height: theme.spacing(2),
-    },
-  });
-
-const useStyles = makeStyles(styles);
+import { css } from "@emotion/react";
 
 function TopSpace() {
-  const classes = useStyles();
-  return <div className={classes.container}></div>;
+  return (
+    <div
+      css={(theme) => css`
+        height: ${theme.spacing(2)}px;
+      `}
+    ></div>
+  );
 }
 
 export default TopSpace;
